refactor(learning): PR-D simplify — secret-leak fix, XSS hardening, file-size split
Three review findings addressed.

1. Secret-leak in detail wire fields (real, security):
   - llm_assessment, materialization, activation blocks were raw-copied to
     wire response without sanitization. Spec layer-6 line 204 forbids
     secret values or redacted-value reconstructions on wire model.
   - New sanitizeAssessmentBlock() + sanitizeProvenanceBlock() helpers
     route every string-valued sub-field through redactObservationText
     before egress. Hash/numeric/boolean fields preserved as-is.
   - 3 new tests in learning-dashboard.test.js (llm_assessment.risk_notes,
     materialization.reviewer_note, activation.target_summary — all with
     API_KEY=<secret> fixtures, asserted absent from wire JSON).

2. XSS via innerHTML (real, security):
   - PR-D added innerHTML usage for 4 detail sections (evidence_summaries,
     llm_assessment, materialization, activation). Strings flowing into
     them could include HTML tags despite secret redaction; the page CSP
     allows 'unsafe-inline' so a malicious <script> would execute.
   - Replaced all 4 sites with createElement + textContent (matches the
     pre-existing card-builder style elsewhere in the file). Pulled out
     an appendLabeledSection helper to dedupe the strong+content pattern.

3. File-size hard-limit overrun (749 → split):
   - learning-dashboard.js was 749 lines; coding-standards.md hard limit
     is 700.
   - Extracted HTTP layer (SECURITY_HEADERS, sendJson/sendHtml/sendError,
     readRequestBody, hasDashboardWriteHeader, createRouter, loadHtml,
     startServer) into new sibling scripts/lib/learning-dashboard-http.js
     (197 lines). Wire-model + action handlers remain in
     learning-dashboard.js (559 lines).
   - Old import paths preserved via re-export from learning-dashboard.js
     to learning-dashboard-http.js so cli.js and existing tests don't
     need to change.

All 1777 Jest + 213 hook + 7 node tests pass; biome lint clean.
