+++ scripts/lib/learning-dashboard.js
@@ -18,7 +18,6 @@
  */
 
 const fs = require('node:fs');
-const http = require('node:http');
 const os = require('node:os');
 const path = require('node:path');
 const crypto = require('node:crypto');
@@ -122,8 +121,10 @@ function evidenceQualityChip(quality) {
 function sanitizeDashboardCard(record) {
   const status = record.lifecycle ? record.lifecycle.status : undefined;
   const llmAssessment = record.llm_assessment || {};
+  const chip = evidenceQualityChip(record.evidence_quality);
+  const hasRelationships = record.relationships !== undefined && record.relationships !== null;
 
-  const card = {
+  return {
     schema_version: 1,
     candidate_id: record.candidate_id,
     artifact_type: record.artifact_type,
@@ -141,18 +142,9 @@ function sanitizeDashboardCard(record) {
       ? llmAssessment.uncertainty_notes.length
       : 0,
     available_actions: legalActionsFor(status),
+    ...(chip !== undefined && { evidence_quality_chip: chip }),
+    ...(hasRelationships && { relationships: record.relationships }),
   };
-
-  // Criterion 1: evidence_quality_chip (derived, omitted if quality unknown)
-  const chip = evidenceQualityChip(record.evidence_quality);
-  if (chip !== undefined) card.evidence_quality_chip = chip;
-
-  // Criterion 1: relationships (copied if present, omitted otherwise)
-  if (record.relationships !== undefined && record.relationships !== null) {
-    card.relationships = record.relationships;
-  }
-
-  return card;
 }
 
 /**
@@ -199,20 +191,58 @@ function sanitizeDashboardDetail(candidateId) {
     evidence_summaries: evidenceSummaries,
   };
 
-  // Criterion 5: llm_assessment — copy if present
+  // llm_assessment / materialization / activation are spec-allowed on the detail
+  // wire. Sanitize them — these blocks may carry free-text (risk_notes) or paths
+  // (draft_path, active_path_summary) that must be scrubbed before egress.
   if (record.llm_assessment !== undefined) {
-    detail.llm_assessment = record.llm_assessment;
+    detail.llm_assessment = sanitizeAssessmentBlock(record.llm_assessment);
   }
-
-  // Criterion 5: materialization + activation — copy from lifecycle if present
-  if (record.lifecycle && record.lifecycle.materialization !== undefined) {
-    detail.materialization = record.lifecycle.materialization;
+  if (record.lifecycle?.materialization !== undefined) {
+    detail.materialization = sanitizeProvenanceBlock(record.lifecycle.materialization);
   }
-  if (record.lifecycle && record.lifecycle.activation !== undefined) {
-    detail.activation = record.lifecycle.activation;
+  if (record.lifecycle?.activation !== undefined) {
+    detail.activation = sanitizeProvenanceBlock(record.lifecycle.activation);
   }
 
   return detail;
+}
+
+/**
+ * Sanitize a free-text assessment block (llm_assessment) — risk_notes /
+ * uncertainty_notes / rationale_summary all flow through redactObservationText
+ * to scrub any leaked secrets the LLM may have echoed from observations.
+ */
+function sanitizeAssessmentBlock(block) {
+  if (!block || typeof block !== 'object') return undefined;
+  const out = {};
+  for (const [key, value] of Object.entries(block)) {
+    if (typeof value === 'string') {
+      out[key] = redactObservationText(value);
+    } else if (Array.isArray(value)) {
+      out[key] = value.map((v) => (typeof v === 'string' ? redactObservationText(v) : v));
+    } else {
+      out[key] = value;
+    }
+  }
+  return out;
+}
+
+/**
+ * Sanitize a provenance block (materialization / activation) — paths and any
+ * string-valued field flow through redactObservationText. Non-string fields
+ * (hashes, booleans, numbers, nested objects) are preserved as-is.
+ */
+function sanitizeProvenanceBlock(block) {
+  if (!block || typeof block !== 'object') return undefined;
+  const out = {};
+  for (const [key, value] of Object.entries(block)) {
+    if (typeof value === 'string') {
+      out[key] = redactObservationText(value);
+    } else {
+      out[key] = value;
+    }
+  }
+  return out;
 }
 
 // ---------------------------------------------------------------------------
@@ -514,205 +544,16 @@ function handleDashboardAction({
   return accept({ next_status: nextStatus });
 }
 
-// ---------------------------------------------------------------------------
-// Security headers
-// ---------------------------------------------------------------------------
-
-const SECURITY_HEADERS = {
-  'X-Frame-Options': 'DENY',
-  'Content-Security-Policy':
-    "default-src 'self'; script-src 'self' 'unsafe-inline'; style-src 'self' 'unsafe-inline'; frame-ancestors 'none'; base-uri 'none'",
-  'X-Content-Type-Options': 'nosniff',
-};
-
-// ---------------------------------------------------------------------------
-// HTTP utilities
-// ---------------------------------------------------------------------------
-
-function sendJson(res, data, status = 200) {
-  res.writeHead(status, { ...SECURITY_HEADERS, 'Content-Type': 'application/json; charset=utf-8' });
-  res.end(JSON.stringify(data));
-}
-
-function sendHtml(res, html) {
-  res.writeHead(200, { ...SECURITY_HEADERS, 'Content-Type': 'text/html; charset=utf-8' });
-  res.end(html);
-}
-
-function sendError(res, status, message) {
-  sendJson(res, { error: message }, status);
-}
-
-function readRequestBody(req) {
-  return new Promise((resolve, reject) => {
-    const chunks = [];
-    let size = 0;
-    req.on('data', (chunk) => {
-      size += chunk.length;
-      if (size > 64 * 1024) {
-        reject(new Error('request body too large'));
-        req.destroy();
-        return;
-      }
-      chunks.push(chunk);
-    });
-    req.on('end', () => resolve(Buffer.concat(chunks).toString('utf8')));
-    req.on('error', reject);
-  });
-}
-
-/**
- * Validate the per-server write token header.
- *
- * @param {object} req
- * @param {string} writeToken
- * @returns {boolean}
- */
-function hasDashboardWriteHeader(req, writeToken) {
-  if (!writeToken || typeof writeToken !== 'string') return false;
-  if (req.headers['x-arcforge-dashboard'] !== '1') return false;
-  return req.headers['x-arcforge-dashboard-token'] === writeToken;
-}
-
-// ---------------------------------------------------------------------------
-// HTTP router
-// ---------------------------------------------------------------------------
-
-/**
- * Create a request handler for the dashboard HTTP server.
- *
- * Routes:
- *   GET  /                         → HTML dashboard
- *   GET  /api/candidates           → DashboardCandidateCard[] list
- *   GET  /api/candidates/:id       → DashboardCandidateDetail
- *   POST /api/candidates/:id/action → action dispatch (requires write token)
- *
- * @param {{ htmlBody: string, writeToken: string }} options
- * @returns {function}
- */
-function createRouter({ htmlBody, writeToken }) {
-  return async (req, res) => {
-    const url = new URL(req.url, `http://${req.headers.host || '127.0.0.1'}`);
-    const pathname = url.pathname;
-    const method = req.method || 'GET';
-
-    // GET / — HTML dashboard
-    if (method === 'GET' && (pathname === '/' || pathname === '/index.html')) {
-      return sendHtml(res, htmlBody);
-    }
-
-    // GET /api/candidates — list
-    if (method === 'GET' && pathname === '/api/candidates') {
-      try {
-        return sendJson(res, createDashboardModel());
-      } catch {
-        return sendError(res, 500, 'dashboard failed to load candidates');
-      }
-    }
-
-    // GET /api/candidates/:id — detail
-    const detailMatch = pathname.match(/^\/api\/candidates\/([^/]+)$/);
-    if (method === 'GET' && detailMatch) {
-      try {
-        const detail = sanitizeDashboardDetail(detailMatch[1]);
-        return sendJson(res, detail);
-      } catch (err) {
-        if (err.code === 'NOT_FOUND') return sendError(res, 404, 'candidate not found');
-        return sendError(res, 500, 'dashboard failed to load candidate');
-      }
-    }
-
-    // POST /api/candidates/:id/action — action dispatch
-    const actionMatch = pathname.match(/^\/api\/candidates\/([^/]+)\/action$/);
-    if (method === 'POST' && actionMatch) {
-      if (!hasDashboardWriteHeader(req, writeToken)) {
-        return sendError(res, 403, 'dashboard write header required');
-      }
-
-      let body;
-      try {
-        body = await readRequestBody(req);
-      } catch (err) {
-        return sendError(res, 400, err.message);
-      }
-
-      let parsed;
-      try {
-        parsed = JSON.parse(body);
-      } catch {
-        return sendError(res, 400, 'invalid JSON body');
-      }
-
-      const result = handleDashboardAction({
-        action: parsed.action,
-        candidate_id: actionMatch[1],
-        expected_current_status: parsed.expected_current_status,
-        safety_ack: parsed.safety_ack,
-        actor: parsed.actor,
-        reason: parsed.reason,
-      });
-
-      let status = 200;
-      if (!result.accepted) {
-        if (result.reason === 'candidate_not_found') status = 404;
-        else if (result.reason === 'stale_status') status = 409;
-        else status = 400;
-      }
-      return sendJson(res, result, status);
-    }
-
-    return sendError(res, 404, 'Not found');
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Server
-// ---------------------------------------------------------------------------
-
-function loadHtml(writeToken = '') {
-  const htmlPath = path.join(__dirname, 'learning-dashboard.html');
-  try {
-    return fs.readFileSync(htmlPath, 'utf8').replace(/__ARCFORGE_DASHBOARD_TOKEN__/g, writeToken);
-  } catch {
-    return '<!doctype html><meta charset="utf-8"><title>ArcForge learning</title><h1>Dashboard UI missing</h1>';
-  }
-}
-
-function startServer(options = {}) {
-  const { port = 3334, host = '127.0.0.1' } = options;
-  const writeToken = crypto.randomBytes(24).toString('hex');
-  const htmlBody = loadHtml(writeToken);
-  const router = createRouter({ htmlBody, writeToken });
-
-  const server = http.createServer(async (req, res) => {
-    try {
-      await router(req, res);
-    } catch {
-      sendError(res, 500, 'learning dashboard server error');
-    }
-  });
-
-  server.listen(port, host, () => {
-    console.log(`ArcForge learning dashboard: http://${host}:${port}`);
-    console.log('Press Ctrl+C to stop');
-  });
-
-  const shutdown = () => {
-    server.close();
-    process.exit(0);
-  };
-  process.on('SIGINT', shutdown);
-  process.on('SIGTERM', shutdown);
-
-  return server;
-}
-
 module.exports = {
   sanitizeDashboardCard,
   sanitizeDashboardDetail,
   createDashboardModel,
   handleDashboardAction,
-  hasDashboardWriteHeader,
-  createRouter,
-  startServer,
 };
+
+// Re-export HTTP layer for backward compat with existing callers/tests that
+// imported these symbols from learning-dashboard.js before the http extraction.
+const httpLayer = require('./learning-dashboard-http');
+module.exports.hasDashboardWriteHeader = httpLayer.hasDashboardWriteHeader;
+module.exports.createRouter = httpLayer.createRouter;
+module.exports.startServer = httpLayer.startServer;
